+++ src/Data/KindID/Internal.hs
@@ -171,6 +171,7 @@ genKindIDs n = fmap KindID <$> V7.genUUIDs n
 nilKindID :: KindID ""
 nilKindID = KindID V7.nil
 {-# INLINE nilKindID #-}
+{-# DEPRECATED nilKindID "Will be removed in the next major release." #-}
 
 -- | Obtain a 'KindID' from a prefix and a 'UUID'.
 decorateKindID :: (ToPrefix prefix, ValidPrefix (PrefixSymbol prefix))